requirements.txt
eventlet>=0.13.0
iso8601>=0.1.9
oslo.config>=1.2.1
pbr>=0.5.21,<1.0
pecan>=0.4.5
six>=1.7.0
stevedore>=0.14
Babel>=0.9.6
Jinja2
oslo.log
oslo.serialization
oslo.utils

